Plush helper methods
